feat: Reflect healthcheck status in exit code (closes #196)
