+++ lib/bin.js
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 // import cluster from 'cluster'
+import { healthcheck, Healthcheck } from './healthcheck.js'
 import { processOptions, cli } from './cli.js'
 
 /* WIP: cluster mode, the main issue is that stores are created by a job before hook
@@ -27,4 +28,9 @@ if (cluster.isMaster) {
 */
 
 const options = await processOptions()
-cli(options.job, options)
+try {
+  cli(options.job, options)
+  process.exit(0)
+} catch (error) {
+  process.exit(1)
+}
