+++ lib/cli.js
@@ -15,7 +15,7 @@ import cron from 'cron'
 import makeDebug from 'debug'
 import * as hooks from './hooks/index.js'
 import { stores, tasks, jobs } from './services/index.js'
-import { Healthcheck } from './utils.js'
+import { healthcheck, Healthcheck } from './healthcheck.js'
 import plugin from './plugin.js'
 
 const { CronJob } = cron
@@ -139,23 +139,12 @@ export function runJob (job, options = {}) {
           else Healthcheck.nbSuccessfulTasks++
         })
         Healthcheck.successRate = Healthcheck.nbSuccessfulTasks / (Healthcheck.nbSuccessfulTasks + Healthcheck.nbFailedTasks)
-        // When not running job continuously stop the server
-        if (options.cron) {
-          return Promise.resolve(tasks)
-        } else {
-          return new Promise((resolve, reject) => {
-            server.close(() => resolve(tasks))
-          })
-        }
+        return Promise.resolve(tasks)
       })
       .catch(error => {
         console.error(error.message)
         Healthcheck.isRunning = false
         Healthcheck.error = error
-        // When not running job continuously stop the server
-        if (!options.cron) {
-          server.close()
-        }
       })
   }
 
@@ -191,7 +180,21 @@ export function runJob (job, options = {}) {
 
 export async function run (job, options = {}) {
   await createApp(job, options)
-  return runJob(job, options)
+  const tasks = await runJob(job, options)
+  // Check if run has to be considered as successful
+  // If CRON the healthcheck command should be used instead
+  if (!options.cron) {
+    let healthcheckError
+    try {
+      await healthcheck(options)
+    } catch (error) {
+      healthcheckError = error
+    }
+    // When not running job continuously stop the server as well
+    await server.close()
+    if (healthcheckError) throw healthcheckError
+    return tasks
+  }
 }
 
 export async function processOptions () {
@@ -200,6 +203,7 @@ export async function processOptions () {
   program
     .version(packageInfo.version)
     .usage('<jobfile> [options]')
+    .allowUnknownOption()
     .option('-d, --debug', 'Verbose output for debugging')
     .option('-a, --api', 'Setup as web app by exposing an API')
     .option('-ap, --api-prefix [prefix]', 'When exposed as an API change the prefix (defaults to /api)', '/api')
@@ -211,6 +215,11 @@ export async function processOptions () {
     .option('-u, --user [user]', 'User name to be used for authentication')
     .option('-p, --password [password]', 'User password to be used for authentication')
     .option('-s, --sync [uri]', 'Activate sync module with given connection URI')
+    .option('-sr, --success-rate [rate]', 'Change the success rate for fault-tolerant jobs to be considered as successful (defaults to 1)', 1)
+    .option('-md, --max-duration [duration]', 'Change the maximum run duration in seconds for fault-tolerant jobs to be considered as failed (defaults to unset)', -1)
+    .option('-sw, --slack-webhook [url]', 'Slack webhook URL to post messages on failure', process.env.SLACK_WEBHOOK_URL)
+    .option('-mt, --message-template [template]', 'Message template used on failure', 'Job <%= jobId %>: <%= error.message %>')
+    .option('-lt, --link-template [template]', 'Link template used on failure', '')
     .parse(process.argv)
 
   // Setup env vars before reading job file so that we can safely use them inside.
