+++ lib/hooks/hooks.utils.js
@@ -3,9 +3,8 @@ import sift from 'sift'
 import _ from 'lodash'
 import common from 'feathers-hooks-common'
 import makeDebug from 'debug'
-import {
-  Healthcheck, callOnHookItems, templateObject, templateQueryObject, transformJsonObject
-} from '../utils.js'
+import { Healthcheck } from '../healthcheck.js'
+import { callOnHookItems, templateObject, templateQueryObject, transformJsonObject } from '../utils.js'
 
 const { getItems } = common
 const debug = makeDebug('krawler:hooks:utils')
