+++ lib/index.js
@@ -7,6 +7,7 @@ export * as services from './services/index.js'
 export * as stores from './stores/index.js'
 export * as jobs from './jobs/index.js'
 export * as tasks from './tasks/index.js'
+export * from './healthcheck.js'
 export * from './cli.js'
 
 export default plugin
