+++ lib/utils.js
@@ -8,18 +8,6 @@ import { Duplex } from 'stream'
 
 const { getItems, replaceItems } = common
 
-// Global structure used to manage healthcheck state for cron jobs
-export const Healthcheck = {
-  isRunning: false, // Flag indicating if job is currently running for cron jobs
-  nbSkippedJobs: 0, // Number of times the scheduled job has been skipped due to on-going one
-  error: null // Indicating error if job has erroned for cron jobs
-  /* Undefined by default
-  nbFailedTasks: 0, // Number of failed/success tasks of last run for fault-tolerant jobs
-  nbSuccessfulTasks: 0,
-  successRate: 0 // Ratio of previous values
-  */
-}
-
 const debug = makeDebug('krawler:utils')
 // Add knot unit not defined by default
 math.createUnit('knot', { definition: '0.514444 m/s', aliases: ['knots', 'kt', 'kts'] })
